requirements.txt
numpy
pyyaml
kafka-python
redis
cassandra-driver